begin drafting updated abstract generator model
