+++ pymoliere/construct/text_util.py
@@ -342,6 +342,7 @@ def analyze_sentences(
         "tok_end": ent.end,
         "cha_start": ent.start_char,
         "cha_end": ent.end_char,
+        "label": ent.label_,
       }
       for ent in doc.ents
       if ent.end - ent.start > 1  # don't want 1-gram ents
