+++ setup.py
@@ -12,6 +12,7 @@ import sys
 proto_src_files = [
     "pymoliere/config/config.proto",
     "pymoliere/query/query.proto",
+    "pymoliere/ml/abstract_generator/sentencepiece.proto"
 ]
